1_data_ingestion/producers/requirements.txt
pandas
confluent-kafka
pyyaml
